% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fspai.R
\name{expandm}
\alias{expandm}
\title{Decompresses a sparse representation of a lower triangular matrix}
\usage{
expandm(Linv, NNarray)
}
\arguments{
\item{Linv}{The Vecchia approximation in compressed format}

\item{NNarray}{A matrix of indices of nonzero entries}
}
\value{
Li The Vecchia approximation in non-compressed format
}
\description{
Decompresses a sparse representation of a lower triangular matrix
}
