% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fspai.R
\name{expandm}
\alias{expandm}
\title{Expand-Matrix}
\usage{
expandm(Linv, NNarray)
}
\arguments{
\item{Linv}{A sparse lower triangular matrix in compressed format}

\item{NNarray}{A matrix of indices of nonzero entries}
}
\value{
The sparse lower triangular matrix in non-compressed format
}
\description{
Decompresses a compressed representation of a sparse lower triangular matrix
}
