% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fspai.R
\name{expandm}
\alias{expandm}
\title{does stuff}
\usage{
expandm(Linv, NNarray)
}
\arguments{
\item{Linv}{Vecchia approximation in compressed format}

\item{NNarray}{Matrix of indices of nonzero entries}
}
\value{
Li  Vecchia approximation in non-compressed format
}
\description{
does stuff
}
