% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fspai.R
\name{FSPAI}
\alias{FSPAI}
\title{Dynamic-Vecchia}
\usage{
FSPAI(covfun, covparms, covmat, NNarray2, locs, m)
}
\arguments{
\item{covfun}{The covariance function used to create covmat}

\item{covparms}{The covariance parameters used to create covmat}

\item{covmat}{The covariance matrix}

\item{NNarray2}{A matrix of indices to search when doing dynamic search}
}
\value{
A list containing
\itemize{
  \item  A matrix of indices of nonzero entries
  \item  The Vecchia Approximation
}
}
\description{
Uses a dynamic algorithm to construct the Vecchia approximation
}
